+++ test/util.ts
@@ -1,11 +1,11 @@
-import "./setup";
-import { expect } from "chai";
-import { util } from "../lib/util";
+import './setup';
+import { expect } from 'chai';
+import { Utils } from '../lib';
 
-describe("util", function () {
-  describe("#chunkedMTU", function () {
-    it("should be 16300", function () {
-      expect(util.chunkedMTU).to.eq(16300);
+describe('util', function () {
+  describe('#chunkedMTU', function () {
+    it('should be 16300', function () {
+      expect(Utils.chunkedMTU).to.eq(16300);
     });
   });
 });